Update the LCD info more often while playing a recording or in live TV.
Currently the LCD info is updated every 60 seconds which means
if you jump to a new position in a recording/live TV the LCD 
could take up to 60 seconds to update the position indicator.

This patch changes the update frequency to once every second which
makes the LCD much more responsive. Uses the unused TV::kLCDTimeout 
const at the top of tv_play.cpp to set what the update frequency 
should be in seconds.


git-svn-id: http://svn.mythtv.org/trunk@8326 4b91aca1-2a1f-0410-a73a-a78f2df169e8