One liner to fix a subtle but nasty bug with oldrecorded.duplicate . If channel surfing in live TV and you happened to flip past a channel that was showing an episode of any show you had previously recorded (whether the showing was a match for a current rule or not), auto-expire would call ForgetHistory() while removing the live file a minute or two later. This would change the status for any matching entries in the oldrecorded table from Recorded to Re-Record by clearing the duplicate flag.
I've known something bad was happening with the duplicate flag for a few
days and fixed a couple related thing along the way. Here is a worst
case scenario:

"Cash Cab" is shown at 6:00 and 6:30 each day with a mix of new and
re-runs. There is a timeslot rule for 6pm. Each showing has a programid
but the subtitle and description for this game show is blank. Several
episodes have been recorded and a mix of new and old are upcoming.

If I flip past this channel at 6:40, a recorded entry would be created
for "Cash Cab" in the LiveTV recgroup. A minute later this would
auto-expire and call ForgetHistory() as auto-expire would always allow
re-record for expired episodes (fixed here by making LiveTV an exception).
Even though there is a programid, the fact that the subtitles and
descriptions were blank would match all showings and mark them all for
Re-Record (fixed in [8313]).  Once the duplicate flags have been
cleared, only the upcoming episodes in the listings could be marked as
"Never record" but none for the previously recorded entries could have
been changed back to Recorded from the Re-Record (fixed in [8313]).

So, if you've been using live TV since the re-design, you may notice
that episodes you've seen before have been scheduled to record again.
This is why. Just mark them as "Never record" when they come up in your
schedule.


git-svn-id: http://svn.mythtv.org/trunk@8358 4b91aca1-2a1f-0410-a73a-a78f2df169e8