References #868.
This is a temporary fix while I work on a more efficient and elegant one.
This fixes the skipframe loop and forces a seek on FF.

The forced seek is the temporary part. Fixing the frame discard will take
me a little time and this should allow people to use single frame seeking
in the meantime.



git-svn-id: http://svn.mythtv.org/trunk@8399 4b91aca1-2a1f-0410-a73a-a78f2df169e8