References #868. References #574.
Fixes a bug with libmpeg2 decoding where we would report a frame as available for viewing before we got a STATE_SLICE with a display_fbuf. This would result in very blocky output after a seek to a keyframe. This fixes the blockyness for low-res PS streams, and reduces it for high-res TS stream. This also releases frames when libmpeg2 is done with them, which should mean less wackyness in low buffer situations (see [8425].)

This bug was documented as a possible problem in the libmpeg2 decoder loop. But it doesn't cause problems during normal playback, only when seeking to I frames, or frames that come shortly after them.



git-svn-id: http://svn.mythtv.org/trunk@8426 4b91aca1-2a1f-0410-a73a-a78f2df169e8