Closes #868.
This is commit is a bit of left over changes from the patch. This refactors
DoRewindDVD() and DoRewindNormal() into a single method DoRewindSeek().

Also there is a small hack in DoFastForward() for MPEG4 NUV files. There
is an off by one error somewhere in NUV seeking, this causes the 
frame-by-frame seek to get stuck in a loop sometimes (one frame in every 
couple hundred it seems.) This hack prevents it from getting stuck in
that loop by simply ignoring one frame rewinds in DoFastForward(). This 
is safe because DoRewind() is only called from DoFastForward() when there
is an off-by-one error, in which case we want the hack, and when we were
decoding normally before the call, in which case this hack is not triggered
because the decoder is more than one frame ahead of the seek position.

There are still problems with seamless ringbuffer switching, but it seems
unrelated to the problems solved in this ticket or caused by [8350].



git-svn-id: http://svn.mythtv.org/trunk@8455 4b91aca1-2a1f-0410-a73a-a78f2df169e8