Fixes a problem where MythTV would crash if we can't find the title to play on a DVD.
We weren't processing the DVDNAV_STOP event, this commit adds some very dumb stop processing. We just stop trying to read data in the DVDRingBuffer after we are not allowed to do it, the only signal higher levels of MythTV get is a failure to read data in DVDRingBufferPriv::safe_read().



git-svn-id: http://svn.mythtv.org/trunk@8472 4b91aca1-2a1f-0410-a73a-a78f2df169e8