Fixes #821.
When flipping through channels quickly enough that the recorder doesn't have time to start up, the recorder is never shut down and so the recording's end time is never set. This sets the end time for these rump 'recordings'.

Also if the recording is less than 30 seconds long the recording end time was often being set as before the recording start time by the rounding rutines in FinishedRecording(). This adds a check for this and sets the end time so that the recording is at least 1 second long in the DB in these cases.



git-svn-id: http://svn.mythtv.org/trunk@8567 4b91aca1-2a1f-0410-a73a-a78f2df169e8