Fixes #981.
The cross recorder channel changes didn't have enough information from
TVRec::CheckChannelPrefix(). This extends that call with some more 
information which allows us to properly switch tuners even if the prefix
matches several channels, but it matches at least one channum exactly.
Before if a channel number was a prefix for another valid channel we
could not switch to that channel unless it was on the current tuner.

Since the change to CheckChannelPrefix changes the MythTV protocol, for
"CHECK_CHANNEL_PREFIX" the MYTH_PROTO_VERSION has been incremented and
a 'make clean' and reinstall of all frontend's and backends is needed.



git-svn-id: http://svn.mythtv.org/trunk@8617 4b91aca1-2a1f-0410-a73a-a78f2df169e8