Closes #919, by applying patch.
The intent of the patch is to make it easier to translate variants of a language. If MythTV can't find a translation in pt_BR, it can pick up the pt_PT string instead. MythContext now returns the language and it's variant when you call GetLanguageAndVariant() and it returns just the language when you call GetLanguage(). I looked over the changed code and grep'ed for other uses of GetLanguage(), but if anyone sees a translation from a different variant than expected anywhere please reopen #919 with an error report.



git-svn-id: http://svn.mythtv.org/trunk@8620 4b91aca1-2a1f-0410-a73a-a78f2df169e8