Closes #989.
This gets a bit closer to the source of the problem than the patch attached to the ticket. OSD::LoadTheme() doesn't like the letterboxing. I've set the bounds to standard ones while LoadTheme runs. They are then reset to the correct ones at the bottom of the constructor (which we were doing anyway.)



git-svn-id: http://svn.mythtv.org/trunk@8644 4b91aca1-2a1f-0410-a73a-a78f2df169e8