Fixes #1053.
In this case an MPEG-4 recording is using a frame height which is not 
a multiple of 16. ffmpeg isn't doing any validation on the dimensions
so it ends up writing outside the buffer. This just pads allocated
buffers up to multiples of 16, so that this type of error doesn't cause
a segfault with any of the codecs we support. MythTV does prevent you
from setting an invalid dimension when you record to a MPEG-4 NUV, so
AFAIK this is more of a legacy issue.



git-svn-id: http://svn.mythtv.org/trunk@8651 4b91aca1-2a1f-0410-a73a-a78f2df169e8