Changes avcodeclock to a QMutex declared in mythcontext.{cpp,h}
The lock was being declared differently in NuppelVideoRecorder and nuppeldecoder, depending on whether you were doing a frontend only build or not. The lock was only being declared properly if you were doing a combined frontend/backend build, and on OS X only if you were using nuppeldecoder, as opposed to say a avformatdecoder.



git-svn-id: http://svn.mythtv.org/trunk@8668 4b91aca1-2a1f-0410-a73a-a78f2df169e8