Fixes #1078. References #956.
The fix for #956 made the invalid assumption that shutting down the
~PlaybackBox() should never take longer than 500 ms to finish. 
This was a zombie prevention measure. I've removed the timeout,
and I've also implemented a deleteLater() function which cleans
up some things early, and should give us snappier preview response.

Torbjorn, can you run with '-v playback' for a while, and then
grep for "previewThreadDone took" in the logs, I'd like some info
on how long this takes in the longest case on your system. On
my system it never takes longer than 7 ms, which is why I thought
a 500 ms timeout would be safe. Maybe I can use a 5 minute 
timeout or something like that..



git-svn-id: http://svn.mythtv.org/trunk@8717 4b91aca1-2a1f-0410-a73a-a78f2df169e8