Fixes #1079, and speeds up scaling.
This grabs the preview from within the output loop, so that it is impossible for a slow scaler/yuv2rgb converter to get parts of two frames. This also speeds up the process 4 to 5 fold by scaling the frame before converting it to RGB, instead of the other way around. This is on my machine with MMX, I wrote the fix with the RGB conversion before the YUV scaling as well for timing purposes and that version is attached to the ticket, so anyone without MMX can test both approaches. 



git-svn-id: http://svn.mythtv.org/trunk@8736 4b91aca1-2a1f-0410-a73a-a78f2df169e8