Fixes #743.
One line patch from Stuart Auchterlonie, for I problem I introduced
when I cleaned up his last patch. Sorry about that Stuart, I should
have sent it back to you for testing since I've never been able to
reproduce this problem reliably.

This ticket is about a error message that can end up spamming the
backend log when the EIT scanner tries to scan a channel that no
exists, or when you have a broken DVB driver, or when you use a
rotor with DVB-S and see other satellites in the sky transmitting
on the same frequency, modulation and polarity as the channel
you are trying to pick up.



git-svn-id: http://svn.mythtv.org/trunk@8788 4b91aca1-2a1f-0410-a73a-a78f2df169e8