References #1147.
* Checks for invalid file permissions in the RingBuffer so we can avoid trying to open a file we will never be able to open.
* Checks if the RingBuffer succeeded in opening the file before starting an NVP. 
MythTV should now tell you if it doesn't have permission to read
the recordings or the previews so you would at least know you had
a configuration problem if you looked at the logs.

The latter missing check may have caused the segfault, but I wasn't able
to reproduce it; the posted segfault is for an earlier revision, and that
particular segfault should aleady be fixed in SVN.

Cougar, can you confirm this fixes the problem for you?



git-svn-id: http://svn.mythtv.org/trunk@8789 4b91aca1-2a1f-0410-a73a-a78f2df169e8