References #1107.
This fixes the hanging part of the problem. The problem is
a LogScale::setMax() uses an iterative algorithm which does not always
converge. This increases the range of the floating point numbers so
that it converges more often, and also limits the number of total
iterations so that the hang never happens.

However, the Visualization is not very informative at higher resolutions.
My guess is that the frequency scale is all wrong since we are increasing
the number of bars but not the size of the FFT window. So I'm leaving the
ticket open in case anyone cares to fix the functionality of the filter...



git-svn-id: http://svn.mythtv.org/trunk@8795 4b91aca1-2a1f-0410-a73a-a78f2df169e8