Closes #1184, by applying Tim Trampedach's patch with some modifications.
This adds an option to the external Firewire channel changing program that allows you to use a single AVC command to change channels. It is always faster than the default method, and is more reliable for at least two people. From what I understand this used to be the default, but was deamed not as reliable as the current default method. However, Tim has changed the retry count and believes that has made it much more reliable.



git-svn-id: http://svn.mythtv.org/trunk@8855 4b91aca1-2a1f-0410-a73a-a78f2df169e8