Fixes #1219.
Streaming requires additional buffering, so it will always be slower to start than using a local or shared directory store.

But this commit addresses an inefficiency introduced by changing the block sizing in [8841]. The FileTransfer had a fixed maximum block size of 256000, but I changed the equivalent block size in RingBuffer to 256 KiB (or 262144 bytes). FileTransfer dealt with this by breaking it up into a 250 KiB and 6 KiB transfer. This didn't make enough of a difference to be noticable on my network (it increases the number of ethernet packets by less than 1%), but on a very marginal system it could make a difference.

I highly recommend using an NFS share or the equivalent for the video directory on HDTV remote frontends. We apply several optimizations for read size and buffering level when using a local RingBuffer which we don't do when streaming. It requires a higher end backend to stream HD material than to shares store with the frontend (faster hard disk mostly); and then it still isn't as quick to start up.



git-svn-id: http://svn.mythtv.org/trunk@8880 4b91aca1-2a1f-0410-a73a-a78f2df169e8