References #1227.
Another fix for PVR-350 decoding. Issac did the coding for this, but asked me to apply it after testing. This is for the frozen input part of the problem reported in the ticket, the RingBuffer readAhead was not being reset after a buffer switch completed so any subsequent buffer switches had the their ignore key mode turned on way too long.

This problem also may have applied to nuppelvideo decoded LiveTV; but there were no reports of this problem with nuppelvideo decoding.



git-svn-id: http://svn.mythtv.org/trunk@8890 4b91aca1-2a1f-0410-a73a-a78f2df169e8