Fixes #1185.
Forces the OSD surface to actually be cleared when the OSD goes away
but the PiP is on screen. When there is no PiP we still just set the
global alpha to zero.

This also changes the paint order so that the PiP is painted before the OSD surfaces, this allows the OSD surfaces to show up on top of the PiP window as intended, rather than under it.



git-svn-id: http://svn.mythtv.org/trunk@8892 4b91aca1-2a1f-0410-a73a-a78f2df169e8