Fixes #1227.
 * This fixes the problem of ateof being set when the reason we couldn't
write data was because the ivtv video decoding buffer was full.
 * This also adds tracking of canwrite, so that if it is true for more than a second during regular playback, we reset playback. That should work around the driver bug where ivtv stops playback if it ever decodes the entire video decoding buffer, even if we immediately refill the buffer.

After the first fix, the canwrite tracking was never triggered for me with 20 second recordings. But you can trigger it if you stick some very short recordings into the livetv chain. This basically simulates the seeks L8r did to get out of the stuck playback with [8883].



git-svn-id: http://svn.mythtv.org/trunk@8893 4b91aca1-2a1f-0410-a73a-a78f2df169e8