References #512. Apply John Harvey's patch for ChromaKey OSD w/XVideo.
This doesn't effect anyone not using this experimental feature. But if you are using the ChromaKey OSD and watch both MPEGs and non-MPEGs this should make the OSD work on those non-MPEGs. This may fail to draw the OSD if you use Xinerama and the XVideo only supports chromakey on a monitor other than the one which the window is on.



git-svn-id: http://svn.mythtv.org/trunk@8906 4b91aca1-2a1f-0410-a73a-a78f2df169e8