Fixes #1105, by applying cleaned up version of funaho AT jurai dt org's patch.
This looks bigger than it is. It touches so many files because it adds a parameter to two audio init rutines. Basically, this is a fix for AC3/DTS decoders that do not auto-detect AC3/DTS using the frame headers. With this commit we open the SPDIF device in the proper mode when starting playback.

Non-auto-detecting decoders will probably not cope well with time-stretch changes; but this commit should now allow the use AC3/DTS passthrough on such devices when time-stretch is not used.

With Mark Spieth's AC3 patches, time-stretch changes should be supported on non-auto-detecting AC3 decoders.



git-svn-id: http://svn.mythtv.org/trunk@8941 4b91aca1-2a1f-0410-a73a-a78f2df169e8