A few commercial flagging changes.
Move the 'All' flagging method up to the top of the dropdown settings box
so that this becomes the new default rather than 'Blank Frame'.  Also fix up
the descriptions in the process.

Fix the help text for the Commercial Skip Method setting to remove the old
outdated comment about not having to have the auto-flagging on when using
blank-frame detection.



git-svn-id: http://svn.mythtv.org/trunk@9002 4b91aca1-2a1f-0410-a73a-a78f2df169e8