References #1133.
Basic support for using both the analog and digital parts of the pcHDTV cards.

Define a DVB card entry for the ATSC portion first in mythtv-setup, then click on the "Analog Options" button and define the NTSC portion.

I'm still working on the configuration UI and testing (esp audio on analog side), but this should already be useful as is.

I'm working on extending this to other dual function cards, but that part isn't ready yet. (You can manually enter the DB entries if you want to try that now).



git-svn-id: http://svn.mythtv.org/trunk@9022 4b91aca1-2a1f-0410-a73a-a78f2df169e8