Add a 'transcoded' flag to the recorded table.  The flag defaults to 0 and is set/unset by the JobQueue for a running, successful, or unsuccessful transcode job.  Support has also been added to the PlaybackBox (aka the Watch Recordings Screen) for displaying a 'transcoded' icon to indicate which recordings have been transcoded.  This is not present in any themes, but the support has been added.
As part of this, I cleaned up the status icon code a bit in playbackbox.cpp
eliminating some repetitive code by creating a loop that checks all icons.
I also added support for the new CC, HDTV, and Stereo icons to the popup
help window since I was working on that area of the code.

This enhances and closes #1308.  I applied the patch in #1308 then added the
frontend support and reworked the status icon code while I was in there.

References #1139.



git-svn-id: http://svn.mythtv.org/trunk@9029 4b91aca1-2a1f-0410-a73a-a78f2df169e8